requirements.txt
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
et_xmlfile==2.0.0
filelock==3.17.0
gdown==5.2.0
idna==3.10
numpy==2.2.4
openpyxl==3.1.5
paho-mqtt==1.5.1
pandas==2.2.3
pymodbus==2.4.0
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
soupsieve==2.6
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
